+++ src/io/ObjValidators.cpp
@@ -82,6 +82,21 @@ namespace xobj {
 	////////////////////////////////////////////////////////////////////////////////////////////////////
 	/**************************************************************************************************/
 
+	bool findHardPolygons(const ObjAbstract & baseObj, const std::string & lodName) {
+		if (baseObj.objType() != eObjectType::OBJ_MESH) {
+			return false;
+		}
+		const ObjMesh * mesh = static_cast<const ObjMesh*>(&baseObj);
+		if (!mesh->pAttr.hard()) {
+			return false;
+		}
+		ULError << lodName << " LOD contains hard polygons in object <" << baseObj.objectName()
+				<< ">. Only the first LOD allowed to contain hard polygons";
+		return true;
+	}
+
+	/***************************************************************************************/
+
 	bool checkParameters(const ObjAbstract & baseObj, const std::string & prefix) {
 		switch (baseObj.objType()) {
 			case OBJ_NO: return true;