+++ doc/changelog.md
@@ -3,6 +3,7 @@ X-Plane Obj Library
 ---------------------------------------------------------------------------
 #### Next
 
+- **Added:** Error if any LOD except the first one contains hard polygons.
 - **Added:** Auto-flip normals for objects with mirrored transformation.
 - **Fixed:** Creating two sided mesh.
 - **Fixed:** Printing custom name of the LIGHT_PARAM.